Add support for long_description_path field in python_distribution (#14448)
Work towards https://github.com/pantsbuild/pants/issues/11554

The rule `determine_setup_kwargs` determines the `kwargs` that are either hardcoded in `provides,` or that are provided by the user's plugin for it. These `SetupKwargs` won't have _everything_, only the foundation for the `FinalizedSetupKwargs`.

We've decided to add the logic of extraction of the long_description file contents here, inside the rule `generate_setup_py_kwargs` that returns `FinalizedSetupKwargs`. That will make sure that this new logic works regardless of if you're using a plugin hook or not.

[ci skip-rust]