jvm: teach dependency inference about multiple resolves (#14491)
As described in https://github.com/pantsbuild/pants/issues/14293, dependency inference does not currently work properly for first-party sources when multiple resolves are configured. This PR fixes the issue for JVM (Java/Scala) first-party sources.

The solution is similar to what was already done for third-party dependency inference: make the resolve part of the key in the symbol map. With this change, dependency inference looks up symbols by both name and resolve.