Properly support filesystem specs in the `paths` goal.
Previously we used UnparsedAddressInputs, which is more
oriented around addresses in BUILD files.

Now we treat the values of --from and --to as specs, so they
can be whatever the user is used to using on the CLI
(as long as they point to a single target, of course).

Also removes the check and exception throwing when there
are no paths, instead emitting an empty list. It isn't an
error to have no paths between two targets.

[ci skip-rust]

[ci skip-build-wheels]