Upgrade Pex to 2.1.66. (#14520)
The changelog is here:
  https://github.com/pantsbuild/pex/releases/tag/v2.1.66

A new setproctitle feature is made transparently available to all
`pex_binary` targets with a direct or indirect dependency on
setproctitle.

The new Pex support for `--complete-platform` will be useful for
dealing with requirements exported by Poetry but is not plumbed in this
change.

(cherry picked from commit 87f457fb00a9a0ba966aff46d937f0cf2494fff1)

[ci skip-rust]

[ci skip-build-wheels]