[internal] BSP: add context type to store connection-related state (#14583)
Introduce `BSPContext` as a way for handler rules to interact with the BSP protocol driver. The first use for `BSPContext` is to store the client parameters sent by the BSP client during initialization. This is then used in the Scala BSP rules to not return any build targets if the client does not support Scala.

Note: Due to limitations in the engine's API regarding what values can be part of a query for a union rule, `BSPContext` is stored in SessionValues. See https://github.com/pantsbuild/pants/issues/12934. We still need to update the `client_params` value after the connection is initialized though, which is accomplished by calling the `initialize_connection` method outside of the engine. Thus, while this class is mutable, it is **not** mutated during calls into the engine. 

Concurrency: There are some concurrency issues with `BSPContext` and the fact that multiple BSP handlers can be active since the protocol allows multiple in-flight requests. See the comments for specifics.

[ci skip-rust]

[ci skip-build-wheels]