Coursier caches are now determined by `-repos` configuration (#14603)
As discovered in #14577, Coursier's maven cache behaviour is to use the local cache before checking remote repositories. This has meant that changing `--coursier-repos` definitions (which should have worked post-#14581) also required a user's pants cache to be wiped (and to not be using any remote caching).

This puts the Coursier maven cache into a subdirectory determined by the values of the `--coursier-repos` argument, which will be observed as part of the Cache key. Coursier fetches should now be properly hermetic.

Closes #14577 again :)