Add `output_path` field to `pyoxidizer_binary` and change its default destination (#14607)
Closes https://github.com/pantsbuild/pants/issues/14565.

Before: `build/aarch64-apple-darwin/debug/install/bin`
After: `src.python/bin/aarch64-apple-darwin/debug/install/bin`

As explained on the ticket, we get rid of `build/` because it was noisy. 

We also need to add a prefix to the directory so that when you have multiple `pyoxidizer_binary` targets, there's no risk of overwriting each other. It's a bummer this causes a longer path, but correctness is more important.

We don't mess with PyOxidizer's platform and {debug,release} folders, as considered in https://github.com/pantsbuild/pants/issues/14565. We decided it's too magical and could cause issues for users who want it.

[ci skip-rust]
[ci skip-build-wheels]