Rename `pants.python.binares` to `pants.core.subsystems.python_bootstrap` (#14613)
It seems unideal to have a top-level folder with only one file. The README for the folder explains it's to isolate code used for finding Python binaries, but there is no reason that needs its own folder, given Pants's file-level understanding of deps.

This adds a new folder `core/subsystems`, where we can define subsystems that are part of Pants's baseline functionality, but which aren't foundational to the engine itself (`pants/engine`). For example, I think it would make sense to move `anonymous_telemetry.py` from `goal/` into `core/subsystems`. This is disruptive so probably not worth it, but we could also move `global_options.py` into here.

### Alternatives considered:

* `pants.backend.python.subsystems` -- this isn't specific to the Python backend.
* `pants.internals.subsystems` -- this would create a new top-level folder. I think it's worth considering, but raises the question of how we decide whether something should be in `core/` vs `internal/`. For example, this `PythonBootstrap` type isn't truly `internal/` - it gets used by the Python backend too.

[ci skip-rust]
[ci skip-build-wheels]