Fix `pyoxidizer_binary` to support `python_distribution` targets that depend on others (cherry-pick of #14620) (#14626)
It's very common for one `python_distribution` to depend on another, e.g. `pantsbuild.pants.testutil` depends on `pantsbuild.pants`. We can easily support that by having the user specify all necessary `python_distribution` targets in the `dependencies` field.

A fancier implementation would calculate for you automatically what dists a `python_distribution` depends on. But that's more complex than it sounds! Note that the dependency is not expressed explicitly via the `dependencies` field, but instead via very complex logic in `setup_py.py`; we'd need to leak that implementation here, or scan the generated `install_requires`. Pants's support for PEP 517 also makes things complicated. So, instead, for now at least, we go with the simple workaround.

[ci skip-rust]
[ci skip-build-wheels]