Export a pip-ified copy of the requirements venv.
Previously we symlinked directly into the named cache,
which was iffy for multiple reasons, including incorrect
interaction with remote execution, and possible corruption
of the cache via that symlink.

Now we create a venv from scratch in the distdir, and
endow it with pip, so users can use it like a regular venv.

[ci skip-rust]

[ci skip-build-wheels]