Give interactive processes a chance to gracefully shutdown (#14580)
## Overview
This PR aims to add a graceful shutdown feature for interactive processes. 

Currently, when `pants run` receives a `SIGINT`, it send a `SIGKILL` to child processes which causes them to shutdown without an opportunity to run cleanup procedures (for instance, `try .. finally` and context managers in python). This can cause file handles to be left open, streams to not be flushed, etc. 

This PR changes the behavior to first send a SIGINT and wait for the process to gracefully terminate within a pre-determined amount of time. If the process does not shut itself down gracefully, a SIGKILL is sent.

This problem is outlined in greater detail in https://github.com/pantsbuild/pants/issues/13230

### Approach
This PR takes approach number (2) from https://github.com/pantsbuild/pants/issues/13230#issuecomment-1028352116. This is was the lowest-lift implementation. It is slightly sub-optimal since it blocks in the `Drop` implementation, however, this should not cause deadlocks since it blocks for a maximum of a bounded time. Furthermore, in my benchmarking, this drop rarely runs, as most of the time the cancellation happens before drop and the drop becomes essentially a noop.

An alternative is to run our own reaping process using a separate routine launched with `tokio::spawn`. The main loop could await this reaping process before shutting down. This would be more complex but would allow us to avoid using blocking calls where we should be awaiting.

Long term a solution for `AsyncDrop` would solve this problem the most elegantly.

Fixes #13230.