go: introduce cache key for compile actions (#14661)
As described in https://github.com/pantsbuild/pants/issues/14625, the Go version and other compiler values that influence the output (other than input file content hashes) were not incorporated into the cache key for building a Go package. (The input files were handled by being part of the input root's `Digest`.) This caused issues when, for example, changing the version of the compiler where incompatible artifacts built by an earlier Go compiler were being used with an upgraded Go compiler.

The `go` tool handles this by incorporating "action IDs" into the "build ID" of a build product. This PR introduces a similar concept to Pants and tries to follow the `go` tool's algorithm as much as possible. (Links to relevant code are in comments in the code.)

[ci skip-rust]

[ci skip-build-wheels]