[internal] Add debug output for intermittent abort in `poll_workunits`. (#14253)
Adds debug output to attempt to catch what we thought was a panic being converted into an abort, but which might just be an all-natural abort from another source.

See https://github.com/PyO3/pyo3/issues/2102 for some discussion of the relevant problem. Beyond just being debug output, this patch actually seems to avoid the issue for the user who was experiencing it.

[ci skip-build-wheels]