Support Java code generation from WSDL files (#14545)
This closes #14523. Still pending for the changes needed to support codegen in compiled languages but the rules for triggering the generation and orchestration of the jvm process that generate the final code shouldn't change much.

This is just a minimal implementation as there could additional enhancements done later (meaning, adding support for more features or customisations).

In the other hand, I'm slightly doubtful of the usability of the _module_ feature (i.e. generate a `module-info.java` file): An scenario in which I see them useful is for cases in which a team/company has developed some SOAP services and uses the WSDL of those services to generate _client SDKs_ that then publishes or shares to other parties. Not too sure if others find this as too niche.