[internal] Implement `Snapshot` in terms of a new `DigestTrie` type (#14654)
As described in #13112: we currently persist _all_ `remexec::Directory` structures to disk, although many of them are intermediate results which never actually need to be persisted across restarts (the only position which actually needs to be persisted is a cache output). That extra IO (and complexity) has performance costs: roughly 20% of some profiles.

In order to skip persisting `remexec::Directory`s to disk, we need an in-memory structure to replace them (which we will persist or load from a `Store` when necessary). To that end, this change introduces a `DigestTrie` type, which is a trie representing a `Digest`ed filesystem tree. A `DigestTrie` is sufficient to replace the list of `PathStats` previously held by a `Snapshot`. But because it uses interned names for path members, and is structure shared using `Arc`s, it does so using a lot less memory.

This is the first of three or four PRs which will replace all methods which recursively load/store `remexec::Directory` with operations on `DigestTrie`s instead (e.g. `MergeDigests`, `DigestSubset`, etc). As methods are ported, their boundaries will be adjusted to skip calling `store.record_digest_tree` (which persists the `DigestTree` to disk using our existing format), until only the local cache persists directory structures.