[internal] Store `DigestTrie`s to local storage using batched writes (#14697)
#13112 is porting more operations on `Digest`/`Snapshot` to use of `DigestTrie` in order to reduce IO by allowing them to operate entirely in memory.

When we have a `DigestTrie` in memory, we are able to persist all of its recursive `remexec::Directory` structures in parallel. To that end, this change adds batch persistence to the local LMDB store, which allows us to use a single context switch and write transaction (*per LMDB shard) to persist an entire `DigestTrie`.