Rename `_python_requirements_file` target to `_generator_sources_helper` (#14727)
Closes https://github.com/pantsbuild/pants/issues/13118 and https://github.com/pantsbuild/pants/issues/12953. 

Really that first ticket was about getting `--changed-since` to work properly for generated targets that depend on the contents of a source file like `go.mod` or `requirements.txt`. (Note that this is different than `python_sources` -> `python_source` etc, which don't actually consume the contents of their `sources` during generation.)

I originally hypothesized that to do that we would need to stop having target generators depend on generated targets. But that was a red herring.

Instead, the simplest solution is to lean into the fix we've used for Python since https://github.com/pantsbuild/pants/pull/9307: generate a distinct target type to "own" the generating sources. Then add a dependency on that target to all the other generated targets. This gives us the exact semantics we want! It's not the most elegant solution—e.g. boilerplate for rule authors—but it solves a real problem well.

[ci skip-rust]
[ci skip-build-wheels]