[internal] Fix Go codegen to work with target generators (#14735)
We used to inject a dependency on the owning `go_mod` target for each `go_third_party_package`. That was causing our compilation code to think that each `go_third_party_package` dependend on _every single target_ generated by that `go_mod`, due to its alias mechanism, which causes issues.

As a result, when you explicitly depend on `protobuf_sources`, we don't replace that with each generated `protobuf_source` target.

Turns out we removed the problematic dependency injection a while ago! So we can fix this easily to properly expand aliases. 

[ci skip-rust]
[ci skip-build-wheels]