Add `[python].lockfile_generator` to allow using Pex for lockfile generation (#14740)
Closes https://github.com/pantsbuild/pants/issues/13964.

A followup will change Pants to use Pex lockfiles for Pants's default tool lockfiles.

## `repository.pex`

Pex allows you to say `pex --lock foo.lock req1 req2`, i.e. install a subset of the lockfile directly.

For now, we stick with our earlier strategy of first installing the entire lockfile, and then using `--repository-pex` to extract subsets later. This greatly simplifies our implementation while we have to support multiple codepaths like `[python].requirement_constraints` - we don't need to change any of that code. It also is necessary for `[python].run_against_entire_lockfile` to work.

We should probably revisit this once we finish stabilizing Pex lockfiles & get rid of the old stuff, based on what will give us the best performance and caching behavior.

## The default generator

For now, we stick with Poetry as the default until we can dogfood Pex more. The hope is in a few days though to change to Pex as the default because of the 3 advantages mentioned in the `help` message:

1) Supports python-repos
2) Avoids the transitive deps environment markers issue.
3) Faster performance when installing.

Co-authored-by: Joshua Cannon <joshdcannon@gmail.com>

[ci skip-rust]
[ci skip-build-wheels]