[internal] New util rule for isolating binary tools. (#14374)
The motivation is to support executing a process as hermetic as possible. Consider that the process requires the use of `ln`, and that it must exist on PATH to work (i.e. it is invoked as `ln` relying on PATH to find it).

If we add `/usr/bin` to the PATH in the sandbox, which would work, we also expose every other binary in `/usr/bin` as well which we don't want to do. (as it defeats the purpose of the sandbox)