Basic support for Helm charts (#14658)
This is a follow up from the massive #14487 PR and the discussion on attempting to split it to make it much easier to review.

The changes in here bring the foundational structures used in the previous PR to be able to handle Helm charts _as a single compilation unit_. To prove that this works fine in combination with the Helm commands, the `lint` goal is also implemented.

Test coverage has been improved in comparison to the same code in the previous PR, but the approach with the `helm_chart` target is the same: Instead of defining fine grained targets, the `helm_chart` target is quite coarse as it includes different types of files as its _sources_: the `Chart.yaml` file, the local `values.yaml` file (which brings the defaults), the `templates/*.tpl` files (go partial templates) and the `templates/*.yaml` files (the kuberentes manifest templates using Go template syntax).

A different approach to this, and maybe one that is worth exploring, would be to have the following targets:

`chart/templates/BUILD`:
```python
helm_partial_templates(name="partials")  # Uses default sources="*.tpl"
helm_kube_tempaltes(name="kube") # Uses default sources="*.yaml"
```

`chart/BUILD`:
```python
 # Has two source fields:
#  - chart="Chart.yaml"
# - defaults="values.yaml"
helm_chart(dependencies=["templates:partials", "templates:kube"])
```

What I like a lot with this approach is that each target is very specific and can be configured very easily. What I mostly don't like is the explosion of targets if produces and having to set the `helm_chart` dependencies expicitly - although this could be mitigated with dependency injection/inference ...

Let me know what you think of current approach VS the one mentioned here.