Target generators (like `python_sources`) now only directly depend on their generated targets (#14766)
Move (rather than copy) `Dependencies` fields from generators to generated targets. This avoids redundant graph edges and calculations for explicit dependencies. And in particular, it prevents us from needing to fill in missing parameters when computing the dependencies of generators (which is challenging, since they no longer have the `Field`s corresponding to their address values for comparison) for #14519.

This has a lot of fallout in tests, which in many cases were using "empty" generator targets (with no sources) as inner graph nodes. Because the target will no longer keep explicitly specified dependencies if it doesn't have sources, those targets either need to gain sources, or become generic `target`s.

To warn users of this behavior change, we additionally deprecate declaring a generator target with an explicitly empty `sources=[]` field.

[ci skip-build-wheels]