[internal] Use a broadcast channel to propagate messages to workunit receivers. (#14807)
Rather than having separate channels with their own locking, allocations, and callsites, this change switches to using a `tokio::sync::broadcast` channel to propagate messages to both receivers simultaneously.

This simplifies the code (with one notable exception: see below), and reduces the number of lock acquisitions on the workunit creation path, which should improve scalability on larger core counts.

The primary downside of the broadcast channel is that it has a fixed, pre-allocated buffer size (because "unbounded channels are bad"... at least for servers. *shrug), and so we need to choose a buffer size that works for all consumers. This increases complexity slightly, because we now warn if the `--streaming-workunits-report-interval` is set too high. In practice though, if that warning ever triggers for a user, we should either lower the default poll interval, or introduce a thread to drain for the poller.

Improves performance of simple usecases by ~2%.

[ci skip-build-wheels]