Deprecate not explicitly setting `[python].lockfile_generator` (#14817)
Our deprecation policy demands that we don't outright change the option from Poetry to Pex. Indeed it is wise to require that. When you upgrade Pants, we don't want that to imply all of your lockfiles will be regenerated into a brand new format the next time you run `generate-lockfiles`. Even though Poetry generation is fairly broken, we shipped it and breaking existing users is a worse sin.

In Pants 2.12, we will be able to change to Pex as the default lockfile generator (but we are not obligated to).

[ci skip-rust]