Remove redundancy in how we apply DEFAULTs in config.
Previously we redundantly did so 3 times - once during
config parsing, and twice during options parsing.
Presumably this was a legacy of the repeated patchworking
of this venerable part of the codebase.

This change removes the expansion of defaults into every
config section, which is unnecessary for both returning defaults
and for interpolation.

It also removes one of the places in which we substituted
the default for missing config values during options parsing.

It leaves just the one required application, where the DEFAULT
values are one of the RankedValues used to select the final
option value.

[ci skip-rust]

[ci skip-build-wheels]