[internal] Use static workunit names, and lazily construct metadata based on the level (#14854)
This change cleans up a few TODOs around workunit storage.

Workunit names are almost entirely `&'static str`, with the notable exception of `NodeKey::Task` names (which are the Python function names of `@rules`). But #14683 began interning `Task` structs, which made `Task` names static as well. This change moves to storing workunit names as `&'static str`.

Additionally, because the `Level` is a field of the `WorkunitMetadata`, the `WorkunitMetadata` was forced to be constructed, even for workunits whose level meant that they would not be rendered. Since the `WorkunitMetadata` contains heavy fields like the string description and user-metadata, this change moves to constructing the `WorkunitMetadata` via lazily evaluated `key=value` arguments to the `in_workunit!` macro (which also has the advantage of being more concise).

Finally: an unused argument to `in_workunit!` is removed.

Speeds up common runs by ~2%.

[ci skip-build-wheels]