FmtResult now operates on Snapshots (#14865)
In order to handle the upcoming change to "unify" both `fmt` and `lint` processes for formatters, we need to have the `message` of `FmtResult` be based on a diff of 2 `Snapshot`s. Since `message` can't participate in `async/await`, we need to change `FmtResult` so it already has those snapshots available (namely `input` and `output`).

This is unfortunately boilerplate for the callers, but that can potentially be cleaned up later a myriad of ways.

This also should be essentially no overhead as `fmt.py` was already doing this query (although it was doing it conditionally. I suspect doing it unconditionally will have very little impact on perf because this is guaranteed to be cached if the output digest differs from the input).

[ci skip-rust]