Adding DigestTrie/Snapshot diffing (#14872)
In order to handle the upcoming change to "unify" both `fmt` and `lint` processes for formatters, we need to have the ability to find what files changed in a formatter rule run. This PR adds a recursive `diff` function to the Rust `DigestTrie`, which is also exposed via the `PySnapshot`/`Snapshot` Python byding/type.

The `diff` method returns a complete set of differences (ours/theirs files/dirs + changed) as this was readily available and could be useful in other contexts. For the purpose of my future changes, I just need the changed files.

Tests have been added Python-side as that is easiest to author/grok.

[ci skip-build-wheels]