Add Protobuf formatting using buf format (Cherry pick of #14907) (#14917)
This PR adds support for Protobuf formatting using the newly released `buf format`. I took the libery of bumping the default version of `buf`since `buf format` requires Buf v1.2.0 and above. Right now there's no check to make sure the user is actually on >v1.2.0, but that shouldn't be too hard to add if needed.

[ci skip-rust]