Write test reports to a standard location under the distdir.
Now users can turn reporting on with --report, without having
to fiddle with the location, but can still modify that location
if necessary. This is similar to how coverage reports work.

Also fixes an issue where we hardcoded `dist/` as the default for
coverage reports, instead of using distdir.

Note that user can, as before, set a custom path outside the
distdir (as they can for coverage reports), but this change makes
it more likely the reports will go to dist/, which is probably
gitignored, instead of to some arbitrary location that may not be.

Basically it makes it easier to turn out reports without creating
complications, and makes use of dist/ more standardized.

[ci skip-rust]

[ci skip-build-wheels]