Store workunits as a DAG rather than a tree (#14856)
As described in #14680: due to memoization, workunits are most accurately represented as a DAG, rather than as a tree.

This PR changes the _storage_ of workunits from a tree to a DAG by giving workunits multiple parents. But it does not yet actually add multiple parents to a workunit, which will be accomplished in a followup change. This portion is worth landing separately because it eases the fix for #14867, while the actual addition of multiple parents is not necessary for that fix.
# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]