Support restricting specific dists to just wheels or just sdists.
This feature is provided in the simplest way that is still useful:

- It only works if you use lockfiles and if they are generated by Pex.
- We don't atttempt to parse this out of existing requirements.txt.
  It must be set in pants.toml.
- The config is global to the repo.

We can expand on this if requested, E.g., if users end up needing to set
this differently per resolve we can add more complex config.
However for now this is a very straightforward way to provide this
sometimes critical feature to users without adding complexity we're
not sure we need.

[ci skip-rust]

[ci skip-build-wheels]