Add support for Helm 3rd party dependencies (#14964)
Providing here the addition of managing Helm 3rd party dependencies (or "artifacts"). Helm charts can have dependencies on other charts, which if located in the same repo, this is already supported as first class dependencies. 

With this PR we add support to configuring OCI registries or Helm "classic" repositories as sources ("remotes") from where we can download pre-packaged Helm artifacts. In a future PR we can add support also for publishing charts into those remotes.

The PR also provides with inference of dependencies of a given Helm chart by analysing the contents of the `Chart.yaml` file. This works for either first party and 3rd party dependencies.

Implementation of "remotes" is heavily inspired by the implementation used in the Docker backend to configure Docker registries. There is a key difference in Helm between registries and "classic repositories": classic repositories need to be configured during the setup phase of the `HelmBinary` as Helm needs to update its local indexes before being able to fetch artifacts from them.

This could potentially be implemented making use of the lockfiles mechanism in Pants, such that if someone adds a new `helm_artifact` or changes any of their settings, then the lockfile needs to be updated which will invalidate the cache for any `HelmBinary` that has been previously configured.

If you think using the lockfiles would be desirable in this context, happy to receive some pointers on how to make use of it and make further changes into this PR.