Add `@rule_helper` decorator, which participates in awaitable collection (#15025)
This PR adds a `@rule_helper` decorator alongside `@rule`. 

It's primary use is to reduce the burden of `@rule` authors having to put all `await`s in their one function (see `pex.py`). The result is that `@rule_helper`s are essentially "inlined" from an engine perspective and do not participate in any of the features of a dedicated `@rule` (memoization, etc...).

It also comes with some guardrails (see the decorator's docstring).

[ci skip-rust]
[ci skip-build-wheels]