Add a workflow for comparing remote cache performance across builds (#15183)
This change adds a script and corresponding manually-triggered Github Actions workflow to compare (remote) cache performance across builds. It was inspired by some infrastructure for release vetting used at Twitter.

The script builds a PEX per build commit (selected by `--build-diffspec(-step)`) and collects the runtimes for a goal across a series of source commits (selected by `--source-diffspec(-step)`). The build commits are the commits actually under comparison, while the source commits are the input dataset to test them on.

Each series of builds runs with only the remote cache enabled, and in a separate cache namespace. This significantly improves reproducibility, but means that the first source commit for each build will always be completely cold.