Cache comparison job fixes: round 2 (#15235)
A bunch of fixes discovered while getting #15183 working against `2.10.x` and `2.11.x`.

The largest change is that each instance of the job now runs for one build commit (against a series of source/input commits). This better matches one of the primary usecases for the job (bisecting to find a regression) and allows for more parallelism.

Note though: this job is fairly limited by the fact that the source commits are in the `pantsbuild/pants` repository, because many of our `BUILD` files, configuration, and lockfiles are implicitly tied to the source version of pants that is being used. The source commit range needs to be chosen carefully in order to be compatible with the binary commit. It's likely that either supporting running against non-`pantsbuild/pants` repositories, or supporting cherry-pick fixup commits should be future work.

<img width="331" alt="cache-comparison-example-inputs" src="https://user-images.githubusercontent.com/46740/164799218-c241f047-972d-44c3-9c8b-9b6f6e272cee.png">