Fix the placement of the pex verbosity args.
The pex subcommands mechanism registers the -v flag twice,
once at the global position and once for each subcommand.
In practice the second registration shadows the first in
the resulting options object.

This may be something to address in pex (the global -v does nothing
so we should probably not register it), but the practical upshot
for Pants is that we have to provide the -v flag after the subcommand
name for it to take effect.

[ci skip-rust]

[ci skip-build-wheels]