[internal] scala: use lockfile for dependency inference parser (#15222)
Use a lockfile for the Scala dependency inference parser when used as a tool (which is how it is primarily used). The Pants repository had a `scala_parser` resolve for development purposes, but that was not used at runtime when the parser was actually invoked for end users.

Rename that resolve to `scala_parser_dev` to avoid any confusion with the `scala-parser` tool resolve and have it share the tool's lockfile. (A comment informs readers that they must keep the requirements in `scala_parser.py` and the `jvm_artifact` targets in the `scala_parser_dev` resolve in sync, or else Pants will probably ask for the lockfile to be regenerated.)

[ci skip-rust]