[internal] java: use lockfile for dependency inference parser (#15223)
Use a lockfile for the Java dependency inference parser when used as a tool (which is how it is primarily used). The Pants repository had a `java_parser` resolve for development purposes, but that was not used at runtime when the parser was actually invoked for end users.

Rename that resolve to `java_parser_dev` to avoid any confusion with the `java-parser` tool resolve and have it share the tool's lockfile. (A comment informs readers that they must keep the requirements in `java_parser.py` and the `jvm_artifact` targets in the `java_parser_dev` resolve in sync, or else Pants will probably ask for the lockfile to be regenerated.)

The `java_parser_launcher.py` rules are merged into `java_parser.py` so that the structure of the Java parser is similar to that of the Scala and Kotlin rules.

[ci skip-rust]